Unable to locate Attribute with the given name [user] on this ManagedType

+++ common/common-dao/src/main/java/com/meysam/common/dao/UserRepository.java
@@ -4,8 +4,10 @@ import com.meysam.common.model.entity.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface UserRepository extends CrudRepository<User,Long> {
+public interface UserRepository extends CrudRepository<User, BigDecimal> {
 
     User findByEmail(String email);
     User findByUsername(String username);
